+++ myapp/__init__.py
@@ -1,5 +1,5 @@
 __title__ = 'myapp'
-__version__ = '0.1.1'
+__version__ = '0.2.0'
 __author__ = 'Alan Crosswell'
 __license__ = 'CCbyncsa4.0'
 __copyright__ = '(c) 2018 The Trustees of Columbia University in the City of New York'